Bugfix: set1 primitive used generic int base type, which caused trouble on some systems

+++ vector/simd/avx2/primitives/create_avx2.h
@@ -48,7 +48,7 @@ namespace vectorlib {
       template< typename U = T, typename std::enable_if< std::is_integral< U >::value, int >::type = 0 >
       MSV_CXX_ATTRIBUTE_FORCE_INLINE
       static typename avx2< v256< U > >::vector_t
-      set1( int a0) {
+      set1( uint64_t a0) {
          TALLY_CREATE_SIMD
          
          return _mm256_set1_epi64x(a0);
@@ -80,7 +80,7 @@ namespace vectorlib {
       template< typename U = T, typename std::enable_if< std::is_integral< U >::value, int >::type = 0 >
       MSV_CXX_ATTRIBUTE_FORCE_INLINE
       static typename avx2< v256< U > >::vector_t
-      set1( int a0) {
+      set1( uint32_t a0) {
          TALLY_CREATE_SIMD
          
          return _mm256_set1_epi32(a0);
@@ -112,7 +112,7 @@ namespace vectorlib {
       template< typename U = T, typename std::enable_if< std::is_integral< U >::value, int >::type = 0 >
       MSV_CXX_ATTRIBUTE_FORCE_INLINE
       static typename avx2< v256< U > >::vector_t
-      set1( int a0) {
+      set1( uint16_t a0) {
          TALLY_CREATE_SIMD
          
          return _mm256_set1_epi16(a0);
@@ -144,7 +144,7 @@ namespace vectorlib {
       template< typename U = T, typename std::enable_if< std::is_integral< U >::value, int >::type = 0 >
       MSV_CXX_ATTRIBUTE_FORCE_INLINE
       static typename avx2< v256< U > >::vector_t
-      set1( int a0) {
+      set1( uint8_t a0) {
          TALLY_CREATE_SIMD
          
          return _mm256_set1_epi8(a0);
